New UserLogin class, UserLoginDAO data access object class, LoginServlet. Much of the login and user session handling changed over from logon class to these.
Created 3 JSP include files for the different parts of the user page - top (search/edit), main links body, and bottom and the user page now references those. Having the include file for the main links body will allow me to reuse that code for the editor screens. I included a parameter for specifying the display mode of the links so they can still be disabled or used for special function within the editor. 

Started a new version of editor screen for much needed improvements.

Added method getDispHtml() to UserLink class.
Added member variable numCats (number of categories) to UserPage class, and added methods to get both that and the number of links.
New MiscUtil class for miscellaneous helper methods unrelated to various objects.